requirements.txt
cloudant
fastapi
pydantic
python-dotenv
typing
uvicorn